requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
distro==1.9.0
Django==4.2.20
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
multidict==6.1.0
numpy==1.24.4
openai==0.28.0
paypalrestsdk==1.13.3
pillow==10.4.0
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
yarl==1.15.2
